[aws][gcp] Simplify storing of account identifiers by removing false abstraction
